02-packer/scripts/requirements.txt
azure-core==1.32.0
azure-cosmos==4.9.0
azure-identity==1.19.0
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
python-dateutil==2.8.1
s3transfer==0.4.2
six==1.16.0
Werkzeug==2.0.1
